Fix splatting an empty array
If we splat an empty array (as shown by the non-test-case path of the
unittest runner changes) then we have -1 args, so a uint would wrap. So,
we use an int16 and then we get -1, meaning 3 changes to 2 and therefor
0 args for the called function.

+++ vmlib/vm.h
@@ -23,7 +23,7 @@ typedef struct
     Value *slots;
     uint8_t handlerCount;
     ExceptionHandler handlerStack[MAX_HANDLER_FRAMES];
-    uint8_t bonusSplatArgCount;
+    int16_t bonusSplatArgCount;
 } CallFrame;
 
 struct _vm
